vdpau: Add object_heap from intel-vaapi-driver
An object_heap works as an allocator. Internally is an array of
buckets. It initializes with a pre-booked allocated memory for
multiple objects that are flagged as free. Allocating an object,
means to flagg the already pre-allocated memory as allocated. If
more space is needed, and all the buckets are filled, then the
heap is expanded pre-allocating more memory.

This structure is useful not only to avoid doing malloc calls on
demand (which may be expensive at real time), but also maps
memory to a given id. This is useful because VDPAU as well as
VAAPI functions return ids for later use.

Issue: OCP_2535