vdpau: Create and fill in config structure
Create the FluVaDriversVdpauConfigObject struct that will hold the
incoming attributes given to create a config object. In contrast,
to what i965 does (that in essence copies the incoming attrib_list
into the struct attrib_list field even if the config attributes do
not match the supported formats), only store the attributes that
are supported, if any.

Issue: OCP_2705