vdpau: Implement vaQuerySurfaceAttributes
Modify vaCreateConfig to not only faily return if the profile is
not supported by the driver but also by the hardware. This change
will allow to cache the supported maximum width and height in the
config object. It is better to cache them, because
vaQuerySurfaceAttributes needs to set them as surface attributes,
and it will be unefficiente to query the decoder capabilities
for everytime this function is called.

Other supported attributes:

  - VASurfaceAttribPixelFormat is added as a gettable
    attribute. The documentation states it can be also settable,
    but I cannot figure out the usecase. In the implementation of
    vaCreateSurfaces, the incoming format argument should be check
    against the supported VASurfaceAttribPixelFormat.

  - VASurfaceAttribMemoryType is added, because in a local test
    this is listed by VA_TRACE logs. However, I cannot figure out
    the usecase of it yet.

Issue: OCP_2706