vdpau: Allow creating context without render targets
According https://github.com/intel/libva/issues/251 it is stated
that num_render_targets can be 0. Our implementation of
vaCreateContext was not considering this. In addition, our
implementation set the context_id to the render targets (a.k.a.
surfaces), and all the rest of virtual functions were checking
that the context was in the surface.

This commit allows to create contexts without render targets and
surfaces without a linked context, except in vaBeginPicture.

The result of this commit is to get surfaces rendered in Chromium.

Issue: OCP_2963